Added routing for Landing page
