+++ src/app/landingPage/landing-page/landing-page.component.html
@@ -11,13 +11,13 @@
     <mat-sidenav #snav [mode]="mobileQuery.matches ? 'over' : 'side'"
                  [fixedInViewport]="mobileQuery.matches" fixedTopGap="56">
       <mat-nav-list>
-        <a mat-list-item routerLink=".">Dashboard<mat-divider></mat-divider></a>
-        <a mat-list-item routerLink=".">Timesheet<mat-divider></mat-divider></a>
+        <a mat-list-item routerLink="./dashboard">Dashboard<mat-divider></mat-divider></a>
+        <a mat-list-item routerLink="./time-sheet">Timesheet<mat-divider></mat-divider></a>
       </mat-nav-list>
     </mat-sidenav>
 
     <mat-sidenav-content>
-      <app-dashboard></app-dashboard>
+      <router-outlet></router-outlet>
     </mat-sidenav-content>
   </mat-sidenav-container>
 </div>
