--- src/app/screens/timeSheet/time-sheet.component.html
@@ -1,3 +1,0 @@
-<p>
-  time-sheet works! Yeah!
-</p>
